Finalize 'registration' component
In this commit, registration component in both backend and frontend
are completed. The registration form is validated on the frontend
as well as on the backend.

Field validation is handled in the backend via hibernate and JPA
annotations and in the frontend via form validation API
provided by angular.

When a user submits her info, a User account is created and
automatic authentication and redirection to requested page occurs,
which gives good user experience.

Exceptions and errors are handled appropriately. Several Exception
classes are created to help giving meaningful messages when
something comes up.