Implement & customize spring social facebook
In this commit, I implemented Spring Social Facebook API
in order to connect to facebook bia Graph API. The problem
is that the current version is not up to date with the
latest Facebook Graph API. The class Album, which has
a field named "coverPhoto" that is the ID of the cover
photo, is of type String. But Facebook Graph API returns
And object named "cover_photo" containing the ID and
the creation time of the photo.

In fact, I added a class named "PhotoReference" to the
source code of Spring Social Facebook API and built
it using GRADLE. The class PhotoReference contains
two fields "Id" and "CreatedTime" so the JSON converter
can map it to the plain result. Then, I changed the field type
of coverPhoto from String to PhotoReference.

The albums and photos are exported to MongoDB before
serving the fronend.

In the front end, I created the facebook connect view
that lets the user connect to his facebook account
if a connection is not found. The list of albums is
then displayed