Save albums/photo to mongodb
In this commit, we save import albums and photos from
facebook when the user first connects to the app.

Pesistance to mongodb is done when the user is connected
to facebook and redirected to the index page.

Fields from Album/Photo to be persisted are selected
based on the needs of our app.

Albums images are saved to disc before persisting
them to their mongodb collection.