Add a request interceptor to ConnectFacebook class
When a user authenticates to the app for the first time
and sends a request to the server to login to facebook
and import his photo albums, the request is intecepted
by a an implementation of ConnectInterceptor interface.
The user email is then saved to the session associated
with the request before the context changes.

After redirection from facebook OAuth system,
the user email is restored from session object
and used to save imported albums to MongoDB
database.

When the angular home component call for the albums,
the same email restored from the current user is
used to form a mongodb query and get a paged list
of albums. The same goes for photos