show the list of photos for a given album
A REST resource has been added in order to list
photos for a given album. In fact, when a user
clicks on an album, a GET request is prepared
and sent to the server where mongodb database
is queried and the list of photos for that
album is gathered as a paged list.

In the next commit, pagination will be added
to facilitate user navigation.