Add angular pagination system
A pagination system is added in the front end
to organize the display of albums and photos.
The implementation of this pagination system
consists of a component containing data inputs
like items total count, current page, page size
and the number of pages to display, as well as
event emitters output carrying callbacks sent
from parent components (e.g. photos.component)

This way, app modulation is ensured.