Integrate angular cli build system with Maven
In this commit, we upgraded angular build system
from SystemJS to the new Angular cli. Fortunatly,
Maven supports the integration of angular
while building backend components.

First, the project is split into three Maven modules:

1) Parent project
2) Backend project
3) Frontend project

The parent module included both the backend and the frontend.
Each module has its own pom.xml file that describs how it's
built.

Because the backend module depends on the frontend, the pom
file contains a reference to the angular project. So, when
the application is built, a jar containing the angular application
is added in the lib folder of the backend jar.

This way Maven takes the responsibility to build the whole
application at once without worrying about compiling angular
components separately.

Also, Node and NPM are installed automatically in the process
of building if they are not found.

Finally, Maven wrapper command is added to the parent project
is order to facilitate the build process for the user.