Install Material Design Bootstrap 4 and Angular 5
in order to use some advanced components from bootstrap 4
in angular 5, I have chosen MDB which makes life easier
in terms of css and jquery components.

The frontend project uses now scss instread of plain css.

Also, I have included the modified version of spring-social-facebook
API which supports the new version of facebook graph API. When
requesting an album from facebook, the graph API returns
an object as cover photo instead of the photo ID. This is
why I've modified and built the old version of spring-social-facebook
in order to be able to benefit from all its power.

With the use of Maven, you can first copy the jar to the Maven local
repository before building the app.

Finally, some modification has been made to the logic responsible for
croping and saving the thumbnail photo to disc.